Create a message.

Create a message.

Usage

envelope(
  to = NULL,
  from = NULL,
  cc = NULL,
  bcc = NULL,
  reply = NULL,
  subject = NULL,
  id = NULL,
  importance = NULL,
  priority = NULL,
  text = NULL,
  html = NULL,
  encrypt = FALSE,
  sign = FALSE,
  public_key = FALSE
)

Arguments

to

See to().

from

See from().

cc

See cc().

bcc

See bcc().

reply

See reply().

subject

See subject().

id

See id().

importance

See importance().

priority

See priority().

text

See text().

html

See html().

encrypt

Whether to encrypt the message. If TRUE then the entire message will be encrypted using the private key of the sender.

sign

Whether to sign the message. If TRUE then the entire message will be signed using the private key of the sender.

public_key

Whether to attach a public key. If TRUE then the public key of the sender will be attached.

Value

A message object.

See also

Examples

# Create an (empty) message object.
#
msg <- envelope()

# Create a complete message object, specifying all available fields.
#
envelope(
  to = "bob@gmail.com",
  from = "craig@gmail.com",
  cc = "alex@gmail.com",
  bcc = "shannon@gmail.com",
  reply = "craig@yahoo.com",
  importance = "high",
  priority = "urgent",
  subject = "Hiya!",
  text = "Hi Bob, how are you?"
)
#> Date:                         Tue, 28 Jan 2025 08:02:52 GMT
#> X-Mailer:                     {emayili}-0.9.2
#> MIME-Version:                 1.0
#> To:                           bob@gmail.com
#> From:                         craig@gmail.com
#> Cc:                           alex@gmail.com
#> Reply-To:                     craig@yahoo.com
#> Subject:                      Hiya!
#> Importance:                   high
#> Priority:                     urgent