Add or query subject of message.

Add or query subject of message.

Usage

subject(
  msg,
  subject = NULL,
  prefix = NA,
  suffix = NA,
  interpolate = TRUE,
  .open = "{{",
  .close = "}}",
  .envir = NULL
)

Arguments

msg

A message object.

subject

A subject for the message.

prefix

A subject prefix.

suffix

A subject suffix.

interpolate

Whether or not to interpolate into input using glue.

.open

The opening delimiter.

.close

The closing delimiter.

.envir

Environment used for glue interpolation. Defaults to parent.frame().

Value

A message object or the subject of the message object (if subject is NULL).

Details

The prefix and suffix can be used to add extra subject abbreviations.

See also

Examples

# Create a message and set the subject
msg <- envelope() %>% subject("Updated report")

# Retrieve the subject for a message
subject(msg)
#> [1] "Updated report"
#> attr(,"class")
#> [1] "encodable"