Parameters for MIME functions

These are parameters which occur commonly across functions for components of a MIME document.

Arguments

content

A string of message content.

disposition

Should the content be displayed inline or as an attachment? Valid options are "inline" and "attachment". If set to NA then will guess appropriate value.

charset

What character set is used. Most often either "UTF-8" or "ISO-8859-1".

encoding

How content is transformed to ASCII. Options are "7bit", "quoted-printable" and "base64". Use NA or NULL for no (or "identity") encoding.

language

Language of content. If FALSE then will not include language field. If TRUE then will attempt to auto-detect language. Otherwise will use the specified language.

description

Description of content.

name

Name used when downloading file.

filename

Path to a file.

boundary

Boundary string.

type

The MIME type of the content.

children

List of child MIME objects.

interpolate

Whether or not to interpolate into input using glue.

.open

The opening delimiter.

.close

The closing delimiter.

.envir

Environment used for glue interpolation. Defaults to parent.frame().