Render Markdown into email

Render either Plain Markdown or R Markdown directly into the body of an email.

If input is a file then it will be interpreted as R Markdown it its extension is either "Rmd" or "Rmarkdown". Otherwise it will be processed as Plain Markdown.

render(
  msg,
  input,
  include_css = TRUE,
  interpolate = TRUE,
  .open = "{{",
  .close = "}}",
  .envir = NULL
)

Arguments

msg
A message object.

input
The input Markdown file to be rendered or a character vector of Markdown text.

include_css
Whether to include rendered CSS.

interpolate
Whether or not to interpolate into input using glue.

.open
The opening delimiter.

.close
The closing delimiter.

.envir
Environment used for glue interpolation. Defaults to parent.frame().

Value

A message object.

Plain Markdown

Plain Markdown is processed with commonmark::markdown_html().

R Markdown

R Markdown is processed with rmarkdown::render().

Regardless of what output type is specified in the input file, render() will always use the "html_document" output format.

Examples

# Plain Markdown

markdown <- "[This](https://www.google.com) is a link."
filename <- "message.md"

# Render from Markdown in character vector.
msg <- envelope() %>% render(markdown)

# Create a file containing Markdown
cat(markdown, file = filename)

# Render from Markdown in file.
msg <- envelope() %>% render(filename)

# Cleanup.
file.remove(filename)

#> [1] TRUE

# R Markdown

filename <- "gh-doc.Rmd"

# Create an Rmd document from template.
rmarkdown::draft(
  filename,
  template = "github_document",
  package = "rmarkdown",
  edit = FALSE
)

# Render from Rmd file.
if (!is.null(rmarkdown::find_pandoc()$dir)) {
  msg <- envelope() %>% render(filename)
}

# Cleanup.
file.remove(filename)

#> [1] TRUE